+++ services/admin/src/routes/dmd/[taskId]/index.svelte
@@ -34,9 +34,9 @@
     isFailedDMDTask,
     isSucceededDMDTask,
   } from "@crkn-rcdr/access-data";
-  import DmdSplitWaitingView from "$lib/components/dmd/DmdSplitWaitingView.svelte";
-  import DmdSplitFailureView from "$lib/components/dmd/DmdSplitFailureView.svelte";
-  import DmdSplitSuccessView from "$lib/components/dmd/DmdSplitSuccessView.svelte";
+  import DmdWaitingView from "$lib/components/dmd/DmdWaitingView.svelte";
+  import DmdFailureView from "$lib/components/dmd/DmdFailureView.svelte";
+  import DmdSuccessView from "$lib/components/dmd/DmdSuccessView.svelte";
 
   /**
    * @type {DMDTask} The dmdtask being displayed by the page.
@@ -52,14 +52,14 @@
   {:else if !dmdTask}
     Loading...
   {:else if isSucceededDMDTask(dmdTask)}
-    <DmdSplitSuccessView {dmdTask} />
+    <DmdSuccessView {dmdTask} />
   {:else if isFailedDMDTask(dmdTask)}
-    <DmdSplitFailureView {dmdTask} message={dmdTask.process["message"]} />
+    <DmdFailureView {dmdTask} message={dmdTask.process["message"]} />
   {:else if isWaitingDMDTask(dmdTask)}
-    <DmdSplitWaitingView {dmdTask} />
+    <DmdWaitingView {dmdTask} />
   {:else}
     <!--JUST In Case All Else Fails-->
-    <DmdSplitFailureView
+    <DmdFailureView
       message="No objects were split from the metadata file."
     />
   {/if}
